+++ include/Types.h
@@ -25,14 +25,14 @@ typedef long long int64; typedef unsigned long long uint64;
 // const:
 const bool SHOW_IN_DETAIL = false;
 const bool PRINT_PRETTY  = true;
-const float MODIFY_TEMPERATURE_DELTA = 0.001;
-const float BOLTZMANN_CONST = 1.38 * pow(10,-23); //BOLTZMANN_CONST = 1.38064852 * pow(10,-23);
-const float INDIVIDUAL_GAS_CONST = 287.05; // [J / (Kg * K)]
-const float TODO_LIMITOR_FACTOR = 1; //0.001;
+const double MODIFY_TEMPERATURE_DELTA = 0.001;
+const double BOLTZMANN_CONST = 1.38 * pow(10,-23); //BOLTZMANN_CONST = 1.38064852 * pow(10,-23);
+const double INDIVIDUAL_GAS_CONST = 287.05; // [J / (Kg * K)]
+const double TODO_LIMITOR_FACTOR = 1; //0.001;
 const int PRINT_PRECISION = 10; // decimal places that are printed in cout
 
 struct GeoCoords {
-    float geoHeight, geoLength, geoWidth;
+    double geoHeight, geoLength, geoWidth;
 };
 
 struct coords {
@@ -40,13 +40,13 @@ struct coords {
 };
 
 struct vector3 {
-    float x, y, z;
+    double x, y, z;
 };
 
 struct airDelta {
     coords newCoords;
-    float moleculesCount;
-    float temperature;
+    double moleculesCount;
+    double temperature;
 };
